requirements.txt
numpy==1.16.2
pandas==0.23.4
feather-format==0.4.0
seaborn==0.9.0
pymc3==3.6
